# Conflicts:
#	target/classes/org/iesvdm/mapper/PedidoMapper.class

+++ src/main/java/org/iesvdm/SpringBootWebMvcJdbcVentasApplication.java
@@ -1,5 +1,6 @@
 package org.iesvdm;
 
+import java.math.BigDecimal;
 import java.util.Optional;
 
 import org.iesvdm.dao.ClienteDAO;
@@ -108,7 +109,7 @@ public class SpringBootWebMvcJdbcVentasApplication implements CommandLineRunner{
 		}
 
 		// Como es un cliente nuevo a persistir, id a 0
-		Comercial comercialNew = new Comercial(0, "Jose M", "Martín", null, 100);
+		Comercial comercialNew = new Comercial(0, "Jose M", "Martín", null, BigDecimal.valueOf(100));
 
 		//create actualiza el id
 		comercialDAO.create(comercialNew);
