+++ src/main/java/org/iesvdm/modelo/Comercial.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -13,6 +15,6 @@ public class Comercial {
 	private String nombre;
 	private String apellido1;
 	private String apellido2;
-	private float comision;
+	private BigDecimal comision;
 	
 }
